Don’t try to read lemmas from wbeditentity response
This code tries to anticipate the possibility that wbeditentity returns
lemmas, but it doesn’t do so correctly: the saveLexeme.lemmas are an
array of {language,value} objects, and the updateLemmas action expects
to be called with such an array, but the lemmas returned by the API
would be an object with language codes as keys, not an array. Remove the
code for now – it’ll be fixed later (T250550).

Also adjust the unit test accordingly, and update the mock responses so
their lemmas match what we’ll eventually expect wbeditentity to return.

Bug: T271105
Change-Id: Ia11ed1df34e09410678eb4e5034a8ae64295a712