Add term text to UnknownLanguage error message
Add a second version of the UnknownLanguage error message which includes
the term text, if it’s available. This will hopefully make the error
message a bit easier to understand for users.

This also declares strict_types in all affected classes, but note that
e.g. $input and $termText are not type hinted: if the request contains
the right (wrong) JSON, then those values might be e.g. booleans or
numbers – we add validation errors in such cases, but don’t abort
validation altogether, so later validation stages have to be prepared to
accept the wrong type for such requests.

Bug: T265783
Change-Id: Idc9f57f425ccc5fdc5c314e15fc826519d668e71