Update lemmas from wbeditentity response
Update the store’s lemmas (and also the language and lexical category,
though we don’t expect those to change, since wbeditentity doesn’t
return the result of patching edit conflicts that might have changed
those two fields server-side in the meantime) from the wbeditentity
response, rather than the data we sent to the API. This means that any
server-side normalizations (trimming whitespace, Unicode things, …) are
applied now.

Note that the header does not yet update itself from the store after
saving, so right now the effect on the lemmas is only visible after
clicking “edit” and “cancel” again. This will be fixed in a separate
commit.

Bug: T250550
Change-Id: I1c2c13945774ebec5820f5b5c679694e34a043dc