Synchronize state with LexemeHeader more consistently
After the save action is finished, the store may have new data in the
lemmas, language and/or lexical category, returned by the API. We should
carry this data into the Vue header object immediately. (Otherwise, it
would only happen on the next call to cancel().)

Also, cancel() should reset not just the lemmas, but also the language
and lexical category from the state. This fixes a minor bug where
editing one of these fields, clicking “cancel”, and then clicking “edit”
again would show the same edited input to the user again, which is
presumably not desired (though it doesn’t seem to have ever been
reported on Phabricator).

Bug: T250550
Change-Id: Ia17be309e837890c8396cf4fc786769aa2cd543b