Update WikibaseRepo::getWikibaseServices() calls
Also, remove the old code that allowed for the WikibaseServices to be a
ServiceContainer – that possibility went away when the code for
repository-based federation was removed, if I understand correctly. At
the same time, add a bit of new code anticipating a future move of the
PrefetchingTermLookup to the service container, which should allow us to
get rid of the reflection code.

Bug: T276931
Bug: T277314
Change-Id: I516338a76a7c4a0d42a65ea73e4f61e426b6bd68
Depends-On: I36f3401f4eb0274bde1de3059a6921991e7c580e