Add unit test to verify empty lists' metadata
The actual fix is in Wikibase, but guarded behind a feature flag that
currently defaults to false; enable it in ResultBuilderLexemeTest and
LexemeGetEntitiesTest. (The version in ResultBuilderLexemeTest is a bit
uglier because that class doesn’t extend MediaWikiIntegrationTestCase,
but I thought manually salvaging $wgWBRepoSettings and resetting the
service is still better than temporarily changing the parent class.)

When the tmpSerializeEmptyListsAsObjects setting defaults to true (or is
removed), three settings lines in LexemeGetEntitiesTest, and the whole
setUp() and tearDown() of ResultBuilderLexemeTest, can be removed again.

Bug: T241422
Change-Id: Ib55945f31a8b534a3096e3c9080abed0b5262c20
Co-Authored-By: sihe <silvan.heintze@wikimedia.de>
Depends-On: Ia2080d8243ee54ca701f17e8cfaeb341705ff0da