Pass context instead of user into newEditEntity()
This is part of a refactoring to ensure that a context is no longer
passed into the MediawikiEditFilterHookRunner constructor. Since the
class needs a context at some point, that context needs to be passed in
elsewhere, and the newEditEntity() method is the “entry point” that is
used in WikibaseLexeme: by calling it with a context instead of a user,
the MediawikiEditEntityFactory can later pass that context, rather than
a user, to (Mediawiki)EditFilterHookRunner::run().

Bug: T280365
Change-Id: Ic776feddd07d082695f983d91797ff88a64c7f01
Depends-On: I875584a1aa0a8a355c625a0e7626d064399848d9