Selenium: Try to wait for replication before reading from API
This doesn't yet touch all tests that edit and then
immediately read from wbgetentities, but only those
where we saw such failures before.
With this applied the tests in question seem to be
more stable and I couldn't observe any more instances
of related problems with them (but that could very
well also be by chance).

Bug: T277862
Change-Id: Ie33d2c67da8a86e705e2577da55c7439cc510781