selenium: Use ChronologyProtector cookie in mwbot
This is an attempt to ensure that requests made via the browser and ones
made directly from Node through mwbot get the same ChronologyProtector
client ID, by getting the cookie from the browser (after making a
trivial database write, to ensure we *have* a cookie) and adding it to
mwbot’s cookie jar. The cookie contains not just the client ID, but also
the database index, but we don’t care about that: the index should also
be stored by MediaWiki in the main stash, we only need to ensure the
client ID is the same between the two environments. (Otherwise, keeping
the cookie in sync between them would be very tedious.)

Bug: T277862
Change-Id: I5a846b7c33b9e94753f8b89fe230e287684c96a9