Selenium: Upgrade WebdriverIO to v7
Update npm packages: @wdio/* and webdriverio.

Patch I4200df03a743ba1022580b56952585e284a68296 has been squashed into
this one to make it actually work.

The video footage of the failed tests made it look like clicking the
"undo" link didn't work. Maybe we somehow have to wait for them to be
"ready"?

Confusingly, `this.undoRevisionLink.waitForDisplayed( { timeout: 20000 }
);` seems to have done the trick, but the same for the restore link
didn't work?

Using browser.pause() here is not nice, but it seems to be the only
thing that reliably works. It would be great if we would find a better
way to this in the future.

Bug: T274579
Change-Id: I119a8eac02468e323a63eb7d75f686c7352e202e