Add $tags to MediaWikiLexemeRedirector::createRedirect() call
The parent class gains a $tags parameter in change Ib2b03a5055, so we
need to pass it through. I added it as a member of the class, rather
than just putting [] directly in the createRedirect() call, because I
assume that the WikibaseLexeme API (wblmergelexemes) will eventually
also want a tags parameter.

Bug: T286778
Change-Id: I7ab20254d9253d133ef9ec37e1bf204f55a3a36d
Depends-On: Ib2b03a50558230f0e7ad17cb79a9d1445926c9d8