Define PREFETCHING_TERM_LOOKUP for all types in client and repo
This was an oversight in Ibfcfaf3ffcdf860c0d36fd6c44c8807dcacf5334 where
not only the entries for "form" and "sense" were missing, but it was
also the wrong file. `WikibaseLexeme.entitytypes.repo.php` is only used
on repo, but PREFETCHING_TERM_LOOKUP is also needed by Lua services and
thus needs to be defined in `WikibaseLexeme.entitytypes.php`.

Bug: T287085
Change-Id: I4c5fbd4a96922583bcf058c48c2f539b94f29a2e