Don’t use array_merge_recursive() to merge entity types
Most of the entity type definitions values are callables or otherwise
individual values. If a key is defined in more than one definitions
file, then array_merge_recursive() will combine those values into lists
of values, which is generally not supported. This can completely break
the wiki if it happens with the wrong key (see T287100).

Instead, we only want a merge that’s one level deeper than a standard
array merge: both of the entitytypes files will have keys 'lexeme',
'form', 'sense', and the values of those keys (which are mappings of
entity type definitions) should be merged, but the values of the
individual mappings should not be merged. We can use wfArrayPlus2d for
this – because array + works the other way around from array_merge()
(when a key is present in both, + uses the first array’s value), put the
repo file first and the general one after it, so that the repo
definitions override the general ones.

Bug: T287139
Change-Id: If4764d0ccfffe36bb6ca249864768531efcf9b9f