Add tags to final Special:MergeLexemes edit
The outcome of this commit is to add the configured tags to the final
edit (the actual merge) made by Special:MergeLexemes (tags for other
edits, moving the data, were already added in change I806c081993).

To achieve this, MediaWikiLexemeRedirector::redirect() needs to pass the
tags into its parent createRedirect() method, even though tags aren’t
part of the LexemeRedirector::redirect() interface method. The same is
true for the “bot edit requested” flag, and here this was previously
achieved by making the flag a constructor parameter of the redirector,
as well as of the service container creating the redirector: hence the
$botEditRequested member of WikibaseLexemeServices. We could’ve done the
same with the tags, but frankly, adding these parameters to the service
container is pretty ugly. Instead, the solution we arrived at is to
introduce a redirector *factory* as a service. The factory contains all
the services needed to create the redirector (revision lookup, etc.),
but receives the bot flag and tags as parameters on its “new redirector”
method, rather than as constructor parameters. (In addition to the bot
flag and tags, the context can also be moved in the same way –
previously WikibaseLexemeServices had hard-coded the global request
context for this.) The service container then only contains this factory
as a service, and the MergeLexemesInteractor gets it injected and uses
it to create a new redirector with the right parameters.

For now, the service container still has the bot flag, because one other
service also needs it in a similar way. We can probably move that to a
factory as well, and then hopefully move the services into the standard
MediaWiki container.

Bug: T286774
Change-Id: I0040c5932dc661f7886ebdec69dff6974336bd4f