Introduce MediaWikiLexemeRepositoryFactory
Like the MediaWikiLexemeRedirectorFactory introduced in change
I0040c5932d, this serves to move the bot flag parameter and context (or
user) out of the WikibaseLexemeServices. Since this removes the last use
of those fields in WikibaseLexemeServices, we can remove them
completely. The methods to create service container instances are kept
for now, but sooner or later we should move the whole thing into
MediaWikiServices (T287945).

Change-Id: I53151db4c631a2f8f50853ac2ca7a5033ae160f1