selenium: Fix failing test when targeting MediaWiki-Docker
`/tests/selenium/specs/special/recentchanges.js` was failing sometimes
(but not every time) when targeting MediaWiki-Docker. The test never
failed when targeting MediaWiki in CI, or in the daily Jenkins job when
targeting the Beta Cluster.

The test fails when looking for text `entrôpi/entropy/ентропы` on
Special:RecentChanges page.

The commit changes the way the last element is found. I've noticed
that the text is located in a `td` tag:

<td data-target-page="Lexeme:L59">

Other recent changes on the page are located in a `td` tag with a
different `data-target-page` attribute value.

<td data-target-page="Item:Q166">

For more information on CSS attribute selectors see:

https://webdriver.io/docs/selectors
https://developer.mozilla.org/en-US/docs/Web/CSS/Attribute_selectors

Bug: T290183
Change-Id: Iaf221f42c0660245757fb136b514342b3d8642d6