Track “all” usage for whole Lexeme instead of Sense/Form
Usage tracking can handle sub-entity IDs, but change dispatching can’t:
when an edit to the page Lexeme:L1 is made, ChangeNotifier /
WikiPageActionEntityChangeFactory only generates a change for L1. That
change will indicate that there are “other” changes, even if nothing in
the lexeme changed apart from its subentities (which is correct), but
this is still only dispatched as a change for L1, and pages that are
subscribed to L1-S1 alone won’t be notified of this change. Therefore,
for now, let’s subscribe to the Lexeme ID instead of the Sense or Form
ID, so that the change is dispatched properly (albeit not efficiently).

The method addAllUsage() is also only called from our trusted Lua code
(it’s not exposed to on-wiki Lua), so let’s add a type hint. (Note that
it was already assuming that the argument is a valid entity ID.)

Bug: T235901
Change-Id: I2c8c590a00b824f83d318c4c80f356906f48c59d