Selenium: ensure we have video recording during login
This test is currently failing during the daily runs in the "beforeEach"
hook. Unfortunately, the video recordings only start when the actual
test starts and so this makes debugging somewhat difficult.

However, since there is only a single test in this test suite, we can
temporarily move the login action into the test itself and so hopefully
get some useful hints on what is going on.

Change-Id: Ifd71e68b546d3a890577cc0b5965f9511d4f4296