Use createMwApp() instead of new Vue()
In preparation for the Vue 3 migration. In the tests, we still need to
use the old style, since createMwApp() is a MediaWiki-specific extension
that’s only available in the vue ResourceLoader module, not when we load
vue from npm in the tests. Effectively, the tests need to partially undo
the migration – add el and potentially store back to the options, and
turn data from a function back into an object.

This slightly changes the DOM layout in some places (mounting now
appends to the element instead of replacing it), so we need to loosen
one selector in the browser tests. (Note that the FORM_WIDGET_SELECTORS
already have `.wikibase-lexeme-form-header .representation-widget`
without the `>` child combinator.)

Bug: T294465
Change-Id: I9d11bfbdce6e2be91847df2bf567bf32b840426b
Depends-On: I6784810fe5f5534682e9c84d730d8bb6fb718f5f