Disable overly flaky browser tests again
These test have been disabled before for T282488 on the REL1_35 branch.
It was thought that an update to wdio6 had fixed the issue, but
apparently, it exists again with wdio7.

It is unclear what the actual cause for this is.  It somehow "fails" to
click the _link_ to the undo page on the Item version history page. Or
it clicks it, but somehow the browser fails to react to it? The video is
both very telling and unhelpful at the same time.  No idea what is the
cause of that. I can't reproduce it locally and it happens only
sometimes in CI.

This affects other repos than Lexeme as well, see for example PS2 in
I6fdaafe5154a18b46703c9e2fe38ca99545b8492

Further commits related to this issue are
https://gerrit.wikimedia.org/r/q/c274cb30ab10461b29bcba6cceecfe4b77546795
on REL1_35
And ba4910e85739d2aa589c6403292f8d7a8d17f589

Change-Id: I58e46c4dc41764c4b80459a353f655f87e7e88fc