Fix LexemeHeader and GlossWidget mounting
In change I9d11bfbdce (commit 4e9499d3df), we changed these to use
createMwApp(), which simulates the Vue 3 mounting API, instead of the
Vue 2 one. One effect of this is that the app replaces the *inner* HTML
of the target element, instead of the *outer* HTML; in other words, it’s
added inside the target element instead of replacing it. I thought that
this wasn’t very significant, but I was wrong: in the GlossWidget, the
resulting double <div class=wikibase-lexeme-sense-glosses> causes a
rather large marging to be applied twice, while in the LexemeHeader we
end up with two <div id=wb-lexeme-header> elements. Fix this by mounting
the elements in a way that more closely resembles the Vue 2 API; the
createDocumentFragment() trick is inspired by Reddit [1] (note: the
function is available since IE6).

The RepresentationWidget is not adjusted, since in this case, as far as
I can tell, it’s not a problem that the target element survives: it has
a different class than the rendered app (you end up with elements like
<div class=form-representations><div class=representation-widget>).

[1]: https://www.reddit.com/r/vuejs/comments/i1vwyw/comment/gk0253i/

Bug: T297328
Change-Id: Ic36fe5027abd02190e672cd66e4cc930088b970c