Remove most of mw.wikibase.lexeme Lua module
Keep splitLexemeId, but remove getLemmas, getLanguage and
getLexicalCategory, which just duplicate the same methods on a lexeme
entity and aren’t really more efficient to call (except for possibly the
small effect that a smaller array/table has to be translated between PHP
and Lua).

While we’re at it, also update some documentation for the entity methods
that was obviously copy+pasted from the functions without making sense
(the entity methods don’t have an ID parameter).

Since this removes a not insignificant amount of boilerplate code in the
library, both on the PHP and Lua sides, leave some small comments behind
as pointers to future developers who might need to reintroduce some of
this boilerplate again.

Bug: T297404
Change-Id: I43f4aa8a32cb46f04c91e56d384fd9e2c3f7915b