Update dependencies
eslint-config-wikimedia is not updated to 0.21.0 because that version
would require touching many of our qunit tests in order to use stricter
assertions. That is best left to another commit.

wdio-mediawiki is not updated to 2.0.0 because that new version requires
the tests written in async mode.

unexpected is not updated because it causes the tests to fail.

grunt-stylelint is not updated because that also triggers an update to
stylelint which is then confused about @-rules and .hyphen()

Change-Id: I777ac4f95203583f074c36892e5ccbaf1540f9f6