Fully migrate to Vue 3
The vue and vuex versions are pinned to what is currently provided by
MediaWiki core.

This patch needs some non-trivial changes to the jasmine tests:
In order to be able to mutate the inEditMode prop, we need to wrap the
component in a wrapper, because of vuejs/core#4874 not being fixed even
a year later. That in turn prevents us from calling the methods defined
on the component under test. Thus we need two functions to create new
widgets depending on which part of the interface we want to test,
methods or props.

Also, this no longer asserts the internal inEditMode value, because 1)
asserting internals is bad practice anyway, and 2) we no longer have
access to that in the wrapped component.

Bug: T304534
Change-Id: I82313a5eb6e8f19088de4a2e831666cdb656b1eb
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>