Add new-lexeme-special-page to Special:NewLexemeAlpha
This adds a submodule for the sources of the JS for the new NewLexeme
page – a Phabricator mirror of the GitHub repository where we develop.
This mirror was set up in T301335 and will hopefully avoid T301273.

However, that Git submodule doesn’t contain build results, and actually
isn’t used directly: instead, we have the build results in this
repository, under resources/special/new-lexeme-dist, and update (copy)
them each time the submodule commit is updated. (Later we might add some
npm command to combine this.)

The special page’s PHP output is replaced with a div with a marker ID,
and a small JavaScript file imports the build results (a function) and
calls them with a config object indicating where the JS UI should mount
itself (on that ID), plus a MessageRepository implementation that
delegates to mw.message(). The ResourceLoader module includes the
existing wikibaselexeme-newlexeme-lemma message, which we reuse for the
new special page; however, we don’t introduce the new message
wikibaselexeme-newlexeme-submit yet (already referenced in the JS),
pending further internal discussion.

Unfortunately, we need not only a ResourceLoader module for our new
JavaScript, but also a “@vue/compat” module, because the built JS
includes a require("@vue/compat") call (as well as a require("vue")
call), and we haven’t found a way to fix that in the build configuration
yet. @vue/compat is a tiny module (with source code inlined into the
resources.php) that just reexports vue under the new name (since
MediaWiki core’s vue module is the compat build). At some point this
module will be removed: hopefully, we’ll be able to use a version of
Wikit compatible with Vue 3 soon; otherwise, at some point MediaWiki
core will switch from the migration build to pure Vue 3, and then we’ll
have to stop externalizing @vue/compat, and instead include it in our
dist files.

Bug: T300882
Change-Id: I774c2cbe1733971699439517ac998eaf09547536
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>