Add npm command to bump resources/special/new-lexeme
This command fetches and checks out the latest main branch of the Git
repository, runs a build, copies the results into the dist folder here,
and finally restores the repository to its previous location.
npm-run-all is used to split the command into smaller steps, though we
still require a Unix-like environment for the `cp` command.

Bug: T300882
Change-Id: Ic1a7f37235147cf119af1e61e8748e2e8e1cb5a4