Let Special:NewLexemeAlpha create lexemes
This adds some backend code to Special:NewLexemeAlpha, mainly copied
from Special:NewLexeme, which allows it to be used for creating lexemes
when JavaScript is not enabled. The form may be further refined in the
future, but it appears to work to some extent already.

Additionally, we pull in a new version of the new-lexeme submodule,
which submits the JS version of the form to the same backend. Since the
JS form uses the same form field names, this also works for creating
lexemes with JS enabled. This is, however, a temporary solution: we
later want to replace this with an API call.

Validation is performed by the field classes; users without JavaScript
will see the errors, but users with JavaScript will just see the JS form
again with no indication why. We probably won’t fix this before
migrating the JS form to use the API instead.

The no-JS version of the form should also be used for supporting IE11
and other non-modern browsers in the future, by replacing the <noscript>
element with its contents (thereby adding them to the DOM). However,
that is not done in this commit.

Bug: T298153
Change-Id: Id9ad7c6e7eb9b2b9b2dba91f9bc66fbb1c8b172f