Add tests for special/new-lexeme to package.json
Test that the new-lexeme-dist/ contents match the git submodule, and
that the commit in the submodule is reachable from its main branch.

This requires at least commit c0734d7890 of the submodule, so that
vue-tsc is no longer part of the build – our CI runs on Node 12, and
vue-tsc requires Node 14 or later. (But the submodule is already on a
descendant of that commit, so it’s fine.)

Bug: T302794
Change-Id: Ibbe283a01e2e75797c2bd4651668bc8af055d897