Properly connect lemma labels with inputs
Using the lemma index for the ID, with no random component. This only
supports a single set of lemmas in a document, but that’s already
inherent in some of the other header markup – the lexeme language and
lexical category also have constant IDs connecting their labels and
inputs – and it keeps the code simpler than making the ID random.

Change-Id: I1e9203c64bae59e2405a2d4e62e7c4eaced8351c