Pass tags into NewLexemeAlpha
With this submodule bump, NewLexemeAlpha expects the passed-in config to
include tags that should be added to the API request to create a lexeme.
(Note that the tags now have to be user-defined tags, whereas in most
other special pages they can technically be software-defined since
they’re only added by the backend; that said, Wikibase’s documentation
has always said that “you also have to create those tags via
Special:Tags”, i.e. that they need to be user-defined.)

The licenseConfig.json file should be renamed, since it no longer only
contains license configuration; I think it makes sense to call it
settings, since the source of the data is also called settings in
Wikibase, not config.

Bug: T302961
Change-Id: I725abc1dc0fdfbef7628810920481525e98a2278