Replace LexemeLanguageNameLookup service with -Factory
In Wikibase, the LanguageNameLookup service is deprecated in favor of
LanguageNameLookupFactory, because the non-factory service depends on
the user language (as the language in which to return language names),
which isn’t always available – in particular, it’s not available in a
ResourceLoader context, so we can’t use it for T306512.

Instead, introduce a LexemeLanguageNameLookupFactory service analogous
to Wikibase’s LanguageNameLookupFactory, and call that with the right
language, as far as possible. (Using it in resources.php will come in a
separate commit.)

Unfortunately, we still have to fall back to the main RequestContext
(i.e. global state) in some cases (though I’d argue that this is no
worse than the previous code, which also used RequestContext::getMain()
in the service wiring). In LexemeViewFactory, we have a Language but no
MessageLocalizer, and I couldn’t find an available class to turn one
into the other (in EntitySchema we implemented this under the name
SpecificLanguageMessageLocalizer, but there’s no such thing in MediaWiki
core). In LemmaLanguageField, we may have an IContextSource in
$this->mParent, but that will only be initialized after we call the
parent constructor, and at that point we’re not allowed to change the
options anymore (HTMLFormField::$mParams isn’t stable to write;
HTMLFormField::getOptions() isn’t stable to override), so it’s a bit of
a chicken-and-egg problem.

Bug: T306512
Change-Id: I57f4574d9ee31c7de0d7426e1b6351651cca02d6