Add language names to ResourceLoader module
The new-lexeme-special-page submodule isn’t ready to be called with
language names yet, so for now just pass in the keys of the object, so
there’s no interface change across that boundary, so to speak.

Put the language names into a separate generated file, so that
settings.json is only based on the settings and nothing else now.

Bug: T306512
Change-Id: I31e5d6d3b94e31f08599a36fc1015bc62ba5832a