Improve languageNameLookup in LexemeViewFactory
Previously, the MessageLocalizer part of the lexeme language name lookup
we got would still rely on the global request language instead of the
language we have. We can avoid that by creating a derivative context
with a custom language, and using that as a context source, so that the
language will be used both for the language code (when creating the
inner/Wikibase language name lookup) and the messages (when looking up
additional languages).

Bug: T306512
Change-Id: I16091887590a491822ad9cc8cb82ee2b710618fe