Show and filter language names in spelling variant lookup
Update the new-lexeme submodule, and pass a Map into it instead of an
array of language codes. (The Map construction should be equivalent to
new Map( Object.entries( languageCodes ) ), but Safari 10 doesn’t have
Object.entries() yet, and we want to support that.)

For the selenium test we need to pick a language code that is more
unique, because the browser test clicks on the first available option
and for the input "en" that is "Old English (ang)".

That "Map" is not defined is probably just an artifact from eslint
working with a very outdated rule-set. We have to figure out how apply
the right rules to the right files in general. We should probably do set
for a subtask of T306486.

Bug: T305436
Change-Id: I98385f0642dae8f68b9e41087ec94151378637a4