Add search message as soon as Special:NewLexemeAlpha loads
This is useful for clients without JavaScript, who should still see some
suggestion to search for existing lexemes.

It’s also useful for clients with JavaScript, so that the link is
already there while the JS code is still loading (and the page doesn’t
move around once the Vue app is rendered). To ensure that the spacing is
the same, also wrap the message in a <p> like in the Vue app.

Bug: T307430
Change-Id: Iea865bbc02234ba30ad2705ca917430c57d45bfc