Add lexical category suggestions to Special:NewLexemeAlpha
This introduces a new configuration variable (which only makes sense on
the repo) with a list of Item IDs to suggest as Lexical Categories
(empty by default). SpecialNewLexemeAlpha.php loads the labels and
descriptions of the listed Items, formats them in a way resembling a
wbsearchentities response, and sends the result to the client as a
config variable. In JS, it’s then available as:

    mw.config.get( 'wblSpecialNewLexemeLexicalCategorySuggestions' )

Bug: T298150
Change-Id: I055f11614171279aaea5ae45a9fc99cff0ced018