Parse all url params on server
All the url parameters are being parsed on the server and sent to the
client as a package, even the ones like the lemma which are not actually
need on the server, but also available on the client. This should make
it both easier to write tests, as this test has to only be written in a
single place. And also it allows us to reuse the constants for the field
names from the SpecialNewLexemeAlpha class.

The alternative was explored in I987f37eeaa99b723ab1596b57053ee4492d49227

Bug: T308117
Change-Id: I65a73b0af5477ccf0da5170098617474957acabd