LexiData: make max Lemma length available as a constant
In order to validate this in the new-lexeme form, we
need access to the maximal allowed Lemma length.
This is currently hard coded in the call in line
155 of WikibaseLexeme.entitytypes.repo.php.
The constant was added to the LemmaTermValidator class
and called as a ReasourceLoader module config param.
The variable injected to the constructor of
LemmaTermValidator was at this time left in its place
to not break tests and other calls. we might want to
seperately refactor it though.

Bug: T308654
Change-Id: I0c945d3ad1c2a412b6b38b4281d08933f89038de