SECURITY: Validate lemma length in Special:NewLexeme(Alpha)
(Note: the SECURITY is meant to mark this as a patch that is initially
deployed to Wikimedia production via the security deployment process,
but the actual security impact is minimal. We just didn’t want the bug
to be abused before it was fixed, hence the non-public deployment.)

The special pages to create a new Lexeme were previously not validating
the length of the lemma. Add this validation, using a very simple
hard-coded validator for now. In parallel, the length limit is being
refactored elsewhere (change I0c945d3ad1); once that change is merged
and this one is made public, we should refactor SpecialNewLexeme(Alpha)
to use that constant instead of hard-coding it, and probably also use a
ValidatorErrorLocalizer as in change I8171bfef73.

Bug: T308659
Change-Id: Id89a9b08e40f075d2d422cafd03668dff3ce7fc9