SECURITY: Validate lemma length in Special:NewLexeme
(Note: the SECURITY is meant to mark this as a patch that is initially
deployed to Wikimedia production via the security deployment process,
but the actual security impact is minimal. We just didn’t want the bug
to be abused before it was fixed, hence the non-public deployment.)

The special page to create a new Lexeme was previously not validating
the length of the lemma. Add this validation, using a very simple
hard-coded validator. (On the master branch, this will be cleaned up in
change Ic4808a3f40, but the release branches can keep the current code.)

Bug: T308659
Change-Id: Id89a9b08e40f075d2d422cafd03668dff3ce7fc9