Split language-name messages to a separate file
This should make the localization of the basic UI
easier for most people. By their nature, all the languages
mentioned here are quite rare and not well-known, so
it's difficult to translate them correctly. They are mixed
with other UI messages, and it gets in the way of the people
who want to complete the localization of the basic UI
into their language.

Change-Id: I925b594b8e1429be4d0fd5e76c5dd850983e92e4