Remove $wgLexemeTalkNamespace option
Always make the Lexeme talk namespace ID the one of the Lexeme
namespace, plus one, as expected by MediaWiki. (Daniel had already
pointed this out in the review comments of change Ieb4f8dd4fd.) In the
documentation, we still mention the effective default talk namespace ID,
so that people can grep for it.

Also, assert that the parameter is well-formed (integer, even, not
negative), and move the arbitrary limit that the namespace ID must be at
least 100 into this assertion as well. Previously, that limit was in the
registerNamespace() function – any namespace ID below 100 would silently
be ignored. This was also pointed out in review for change Ieb4f8dd4fd,
and seems to be completely arbitrary (dating back to the initial commit,
based on the Boilerplate extension), but I don’t think it’s worth
actually *changing* at this point.

Also, document the behavior that the setting can be set to false to
disable registering a namespace. I’m not completely sure how this is
supposed to be used, but if it exists, let’s document it, I guess.

Bug: T308549
Change-Id: I1e6bb88c27696a6e68323658bb59fa8a770751b2