Clean up lemma length validation
This follows up change Id89a9b08e4 (commit 46ae8fb7e8), which introduced
the missing validation with some hackier code – the constant for the
LemmaTermValidator::LEMMA_MAX_LENGTH was being introduced around the
same time, so directly using it in the security patch would have added
some awkward cross-dependencies between public and private patches.

The cleanup has two main parts: move the LemmaTermValidator to the
service container, which also benefits WikibaseLexeme.entitytypes.repo;
and then use that, as well as a ValidatorErrorLocalizer (a Wikibase
service), in the special pages.

Slightly annoyingly, an HTMLFormField validation-callback isn’t allowed
to return a Status, so rather than simply returning

    $this->validatorErrorLocalizer->getResultStatus( $result )

without any conditional (the Status is either good or not), we have to
check ourselves if the result is valid or not, and manually pick out the
first error (of potentially several) to get its error message. Oh well.

Bug: T308659
Change-Id: Ic4808a3f40ca8b253233d8996fe0f76fc6afee99