Use FallbackLabelDescriptionLookupFactory in LexemeHandler
This makes no difference whatsoever, because the resulting lookup is
only used to get the label of the entity whose history we’re looking at,
and Lexemes don’t have labels.

In the test, the mocked return value for newLabelDescriptionLookup() can
be removed – the method has a static return type, so PHPUnit can return
a mock of the correct class by itself.

Bug: T312223
Depends-On: Ia94c239869bc6525b70039030920ae7fe2f3be90
Change-Id: Icbe3b235d487c4b1bc581a96a7d9bdd8578e6e8d