Use FallbackLabelDescriptionLookupFactory in SpecialNewLexemeAlpha
This resolves redirects in the example lexeme’s language and lexical
category when they are used in the placeholders of the corresponding
form fields. Note that the links to those items in the info panel still
don’t resolve redirects, since those are rendered differently – I think
the fix for that will be in Wikibase.

It is also used when the items are loaded from the URL parameters –
compare the previous change (change I7bab4e11ca), which fixed this case
for Special:NewLexeme (and no-JS in NewLexemeAlpha).

Bug: T312223
Bug: T312658
Change-Id: Iab119ee946b76d82f6c592cd3efe9d939947bae4