Use FallbackLabelDescriptionLookupFactory in SpecialNewLexeme
This is used when the language or lexical category have been provided
via the URL. If a merged item is selected, the label of the target item
will now be shown.

Also, add a TODO comment about the strange choice to hard-code the label
language to English, which shouldn’t be fixed in this change.

Bug: T312223
Bug: T312658
Change-Id: I7bab4e11ca7c8f17bf306e4fb94f03bc135a1efd