Lexemes and Forms don't have Labels, so don't try to look them up
This reuses the NullLabelDescriptionLookup that was introduced with
Forms in 2018: I675b39bf (commit cbf31e2).

The Wikibase classes EntityIdSearchHelper and HistoryEntityAction couple
the concepts of "Entity" and "has Terms", even though they are distinct.
We should probably refactor them at some point to be more agnostic. But
for now, explicitly using a NullLookup seems more transparent as to what
 is happening compared to trying to lookup non-existing Terms.

Change-Id: I0ffbe08ffa4dfd00c5d59261d9f7fff1c72cdd96