Make language fallbacks work for example lexeme ID
If a message exists in the CDB cache (built from the JSON files),
MediaWiki will prefer the CDB cache version (even just the final English
fallback) over any on-wiki fallback (other than a page for the exact
target language), which means that the example lexeme ended up being L1
in most languages. The solution is simple: remove the message from the
CDB cache. We don’t need it to exist in the JSON files or on
translatewiki.net, since it’s not supposed to be translated anyways; and
the final fallback to L1 can be hardcoded into SpecialNewLexemeAlpha.php
instead, if the message doesn’t exist.

With this change, language fallbacks between different MediaWiki: pages
for this message are fully functional (as far as I can tell), so add
tests for this behavior: make the existing test check for a final
fallback to en, and add another test (based on the first one, but
without the HTML-escaping checks) for a regular fallback (de-at → de).

Bug: T313599
Change-Id: I93f9b645eaf513ee482ea1bedcd1b425790e6b73