Allow removing terms with invalid language codes
This is necessary to let users remove *-x-qid terms once we disallow
that in favor of *-x-Qid (including to let users replace *-x-qid with
*-x-Qid, which is technically a removal + an addition).

To implement this, LexemeTermSerializationValidator::validate() is split
into two methods, validateStructure() and validateLanguage(), and
callers should only call the latter method if validateStructure() passed
and the term isn’t being removed. It would be possible to keep this in
one method, and have validate() skip the language part if it determines
the term is being removed, but that would mean that this check (is
'remove' set or 'value' empty) would be duplicated between the validator
and its callers, which doesn’t sound great to me.

Bug: T317863
Change-Id: I57867866c7c902b23f367b1c4ede39369c44db6f