Only send edited lemmas when editing lexeme header
This should improve performance a little bit, but more importantly it
avoids validating the lemmas that weren’t edited. If a user is editing
the en lemma, we don’t want to block the edit just because another lemma
that also happens to exist uses a language code that is no longer valid
(e.g. en-x-q123 instead of en-x-Q123). Note that we already do the same
thing for form representations (resources/entityChangers/FormChanger.js,
getRepresentationDataForApiRequest()).

In the tests, add more test cases for editing lemmas, including some
where the behavior is now different. Also, make sure to call done.fail()
when an error occurs, so that we can see the error details (a
pretty-printed diff of expected and actual value) instead of a timeout.

Bug: T317863
Change-Id: I1e9369de4eda71ca2a22980cfd0f86617ad7ee6e