Inherit skin font family on Special:NewLexemeAlpha
We want the special page to use the same font as the rest of the wiki
interface. Because the skins don’t expose the font family in a uniform
way (e.g. Vector only sets it on html and body, with very low
specificity), we couldn’t find an elegant way to do this; an attempt to
fix this in Wikit ultimately had to be abandoned.

This solution isn’t very elegant, but it does the job. All the CSS rules
in resources/special/new-lexeme-dist/style.css that include a
font-family declaration have a selector with specificity 0-2-0 – either
two classes (.wikit .wikit-Button, .wikit-Button.wikit) or a class and
an attribute selector (.wbl-snl-copyright[data-v-6ba2e1db]). We can thus
override them with a specificity 0-3-0 selector, by repeating the same
class three times, or a specificy 1-0-0 selector using an ID; by
unsetting the font-family on all the matched elements, we let them
inherit the font-family from the surrounding elements instead, where the
skin set the font-family.

Bug: T313166
Change-Id: Id60cca3b941b46885ca77ee485fba49cb9f173e2